fix: Phase 0 — Materialize error propagation, Version.Add(uint), V3_MIGRATION
Three critical bug fixes:

1. stack/materialize.go: Any Store.Get error was silently classified as
   "not found" and triggered OnCreate, causing silent data corruption on
   real DB errors (connection drops, decode failures). Now uses
   errors.Is(err, kv.ErrNotFound) and propagates all other errors.

2. event/types.go: Version.Add(n int) accepted negative values which
   silently wrapped to huge uint64 values after the int→uint64 change.
   Changed to Add(n uint) — the type system now prevents negative input
   at compile time. Updated all callers (batch.go, decider.go, load.go).
   Also fixed storage/pg_bus_listen.go np.Version-1 underflow guard.

3. docs/migration/V3_MIGRATION.md: Fixed broken memory.NewMemoryBus()
   example (module already deleted). Corrected ghost-code lie:
   storage/pg_bus.go is LIVE code (ADR-0027), not ghost. Updated status
   table to show Version→uint64, memory/ move, and readmodel deletion as
   Done; io.Closer removal as Rejected (verschlimmbessern).

Added test: TestMaterialize_StoreGetErrorPropagates verifies that
non-ErrNotFound errors propagate instead of triggering OnCreate.

+++ decider/decider.go
@@ -156,7 +156,7 @@ func (r *Repository[State]) Execute(
 		}
 	}
 
-	newVersion := currentVersion.Add(len(newEvents))
+	newVersion := currentVersion.Add(uint(len(newEvents)))
 
 	r.saveSnapshotAfterEvents(ctx, ref, newVersion, state, newEvents)
